Replace the current vout by a native surface vout
This commit merges the vout module made by Ming Hu in his github project
and removes the current vmem/opengl code.

This new vout is much faster than the previous: It is able to draw and
scale a 720p-h264 video with only few frame drops on a Nexus One.
The counterpart is that it uses some non-public and non-documented APIs.

Tested on a Nexus One running Android 2.3.4 and a Xoom tablet running
Android 3.0 (Honeycomb).