Fix an IllegalStateException related to the onSaveInstanceState call
The error appears because we're attaching a new fragment in the onResume
and this method is sometimes called after onSaveInstanceState which
leads to a crash.
We can safely discard the state in this specific case because we don't
rely on the UI state but on the mMediaLibraryActive boolean to read the
actual state of the directory view.