Move the ThumbnailerManager into the MainActivity
Generating the thumbnails being a background job of the MainActivity
(the same way that the file scanning of the media library) it makes
sense to be managed by the latter and not by one of its fragment.

Furthermore the ThumbnailerManager is now a singleton so it can be
accessed indifferently from the MainActivity or the VideoListFragment.