VideoGridFragment: determine at runtime the best viewing mode
    Because of a broken GridView API we're not able to get the number of
columns the GridView is able to show without laying it out first. And
reconfiguring it to switch from list to grid (and vice versa) is also a
pain in the ass because of the poorly designed getNumColumns method.
    Therefore we are forced to compute manually, using constants
(because on Android you can't query the XML), the needed size to show at
least two columns. Otherwise we have switch to the list mode.