Add an LRUCache to store the frequently used bitmaps
The cache can be used to store almost any kind of bitmaps but its main
purpose is to speedup the UI (like lists and grids) and because the
memory of the cache is limited and constrained by device your should
avoid storing bitmaps that are not going to be shown very often or it
would void the benefits from the system on low-memory devices.

If you think this sentence is very long, you never read Proust.