Rework of the remote control handling (for wired, bluetooth and lock screen)
Now VLC doesn't steal the remote control to other media players but the
selection is made by the system based on the last application which
requested it.

Except that it doesn't introduce any other functionnal change in the
behavior (double click should still do a forward and long click a
backward).

This commit also merges the two broadcast receivers into a single one to
avoid code duplication although a double click on the play/pause button
of the lock screen will trigger a next. Not a big deal I think.

Support summary:
On API level 7: use the old method (global receiver)
On API level 8+: correctly share the remote with other media players
On API level 14+: support for the lock screen controls