Show 'sdcard' instead of the user's ID in the file browsers
Since Android 4.2 the support for multi-user accounts was added and uses
a separated storage space for each user represented by an ID
(/storage/emulated/<user-id>) that is visible in the file browsers.
Instead of showing this nonsensical number, and since we're unable to
get the current user name (thanks to the protected state of the
UserManager class), this commit replaces it by 'sdcard' to mimic the old
behavior.