Revert "Set SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION in the video player"
This reverts commit a12edf9d192b79f3b8fa0044395bf66ae8afc5db and
a8822b341483da26338926058831c2076e2377cd.

If this flag is set, the layout needs to be adjusted to make sure
that the vlc controls aren't laid out under the navigation bar.
That is, we'd want the effect of this flag on the video surface,
but not on the rest of the video player UI.

Signed-off-by: Martin Storsjö <martin@martin.st>