Return to the MainActivity after the AudioPlayerActivity finishes
When the player is started from a notification (without a backstack) and
it is closed (by calling stop or when the end is reached), the
MainActivity will now always be shown.

In the same way, returning from the application history will launch the
player if it's already running instead of the menu.