VideoGridFragment: Call updateList later at startup
Don't call updateList() in onViewCreated, call it in onResume
just before registering the view as a listener to updates.

This reduces the probability for race conditions. Earlier, this
was pretty often visible on the emulator, but also occasionally
on real devices. If the MediaLibrary update finished after
the grid view fragment's onViewCreated method, but before onResume,
the fragment wasn't ever registered for receiving updates about
the media library, and had only updated the list while it was empty.

Technically I guess there's still a risk for race conditions here,
but by moving the updateList() call closer to the addUpdateHandler()
call (and at least calling it in the same FragmentActivity callback)
at least reduces the risk.

Signed-off-by: Martin Storsjö <martin@martin.st>