Keep a stable layout when hiding the navigation bar
By adding the SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION and
SYSTEM_UI_FLAG_LAYOUT_STABLE flags, the activity can be laid out as if
the navigation bar was hidden even when it isn't, avoiding moving the
video when the UI disappears.

This requires setting the fitsSystemWindows="true" parameter for views
that still should be laid out as if the navigation bar was there. That
is, the video surface is laid out to cover the whole screen (under the
navigation bar), while the UI on top gets padding enough to not be
obscured by the navigation bar.

Signed-off-by: Martin Storsjö <martin@martin.st>