widget: fix widget updates when launcher is restarted
When the launcher is killed/restarted, only the latest RemoteViews is restored.
so use partial updates if possible (API11+), else update everything.

Unfortunately, in API below 17, partial updates are not cached by android,
so we have to force those updates from the service.

To avoid spamming with several big intents during playback,
the cover is not restored, only state, text and position.