libvlc: Don't try to load the iomx library on kitkat
MediaCodec should be preferred over iomx, since the former is a
public API, and is always availble on kitkat. The IOMX ABI has
changed between jb_mr2 (which used the -ics binary) and kitkat,
and trying to load the -ics binary will actually succeed, but
will crash if one tries to use it.

The VLC MediaCodec module currently only supports video but not
audio, but there's seldom any point in using IOMX for audio
anyway, and if this actually is an issue, the MediaCodec module
can be updated to support audio decoding as well.

Signed-off-by: Martin Storsjö <martin@martin.st>