VideoPlayerActivity: hack to fix lost overlay controls
DVDs are not the only type of media that can contain titles and chapters.

But most other media only have one title (e.g. MKVs) unlike DVDs, so don't hide the overlay for normal files.

A better solution is a button to simply select the title/chapter desired and would work with both DVD and normal media.