Update the hardware/libhardware headers for honeycomb
Previously this used a checkout from ICS (which worked fine so far).
There are no honeycomb tags in this repo, so we can only guess and
pick commits based on where important changes were made.

This set of headers is from commit 41693fa, which is the last commit
before gralloc.h started including system/window.h (which caused issues
with the system/core headers from ICS, which provided a duplicate
definition of android_native_buffer_t.

Signed-off-by: Martin Storsjö <martin@martin.st>