Replace SlidingMenu lib by support MenuDrawer
MenuDrawer is the menu framework provided by Google in android support library v4.
This removes a dependency on a 3rd party lib.

A remaining issue is that the menu listview is wiped from main layout when audio player is shown,
I had to re-add it once SlidingPanel is opened as a work-around.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>