Save fragments to reuse them
To not destroy fragments, I add them to the backstack, but I don't use it.
Only to retrieve fragments by tag.
That's why we finish() on back pressed if no other action is planned.

I don't see any other solution for now, documentation explains that fragments not in backstack are destroyed once replaced.
An alternative could be to change the way fragments are handled, and don't use ft.replace() anymore.

After a few testings it seems to work pretty well, and no extra memory is needed