Always Open stream from VideoPlayerActivity
And fallback to AudioService if there is no video.

Opening Video from AudioService can't work with the new android_window vout
module because the module need VideoPlayerActivity to be created at open.
Indeed this activity won't be created if the event MediaPlayerVout is not
received, and this event won't be sent if android_window is not opened.