MediaWrapper: remove constructor with LibVLC
This constructor was parsing a media and could take a long time. Replace it
with a mrl constructor that don't parse a media.

Indeed, MediaWrapper is only a wrapper and is often used in the UI Thread. It
shouldn't do any parsing (but it still fetch data that is already parsed when
constructed from a media).

Add "// FIXME" for all media that are parsed in UI Thread.