Only one way to get LibVLC instance.
There were 2 differents way to get a LibVLC instance. For now, use only
VLCInstance.get() utility method. This method will init LibVLC if it's not
already done.

This method will throw an IllegalStateException if LibVLC.init() failed. It's
not a compatible CPU error, the application will crash with the default android
crash dialog instead of the custom CompatErrorActivity. It allows the error to
be reported to the google play dev console.