libvlc: add std_logger.c
By default, stdout/stderr are deactivated, i.e. they are opened on /dev/null.
std_logger_Open will duplicate stdout and stderr fd on a write end of a pipe. A
thread will be created and will print the read end of the pipe.

Standard output will be printed with "VLC-std" TAG with debug priority.
Standard error will be printed with "VLC-std" TAG with error priority.

It's not activated by default since we don't want to mess up with the
application file descriptors.