Fixed inappropriate use of Calendar.roll in TimePickerDialogFragment
This fixes a bug that caused playback to stop immediately when it is
the last day of a month and you set the sleep-timer after midnight. The use
of Calendar.roll caused the resulting target-date to be in the past.
Calendar.add should be used instead of Calendar.roll so larger fields get
rolled over too.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>