minSDK is now 8, aka Android 2.2
This is mostly because of the AudioManager changes, but also the fact
that 2.1 is quite old and rare.
The Play Store is also only 2.2+ since a long time.

However, developers, please try to keep your work compatible with Android-7
as much as possible. We don't know what the future will be.