compile.sh: error on invalid usage
If a user of the script happens to pass an invalid option to
"compile.sh" we should not ignore it, instead it is more appropriate
to add a diagnostic saying that the user has supplied an invalid
option.

As such, these changes adds a fallback entry to the argument parsing
so that the script aborts to save user from unexpected behavior.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>