compile: robustly initialize local.properties.
If local.properties exists but has been modified since last build,
it might need to be changed to contain appropriate android sdk/ndk paths.
For example, Android Studio will automatically modify local.properties
if the project is opened in an environment with a different sdk location.

Supports the following operations:
 - create a new local.properties if none exists, same as previous commit.
 - append sdk.dir and ndk.dir lines if none are found in existing file.
 - if existing file has incorrect sdk/ndk values, or has multiple entries
 for either, those lines will be replaced with a single one for each of
 sdk.dir and ndk.dir.

Signed-off-by: Geoffrey Métais <geoffrey.metais@gmail.com>