Clear list filter on small search query string
When a query having less than 3 chars was requested, it did nothing.

As a consequence, searching "xy" did not always give the same filtering:

   query   filter applied
   -----   ----------------
      ""   none (full list)
    "xy"   none (full list)
   "xyz"   "xyz"
    "xy"   "xyz"            <= unexpected

For consistency, always display the full list when the search query is
small.

This paves the way to keep the filter after an activity is destroyed
then re-created (e.g. on screen rotation).

Signed-off-by: Geoffrey Métais <geoffrey.metais@gmail.com>