build: force failed detection of search.h and tdestroy
Fixes build with the NDK 17

Force failed detection of search.h in order to avoid conflicts with vlc_fixups.

Force failed detection of tdestroy so that fixups will declare it. Since tfind
is always available (found) tdestroy compat function will never be built (which
is the wanted behaviour since it is always available too).