build: restore c++_shared usage
This is a partial revert of the following commits
 - b775dbba3acdfd78adb5fb0a45bd607df7ea0f3e "build: fix ODR violation"
 - b9df53313bae2fffb9622d725060bd3f6fc5d68a "Use libc++ as static library"

UnsatisfiedLinkError with c++_shared should be fixed since we now use the last
NDK version and we dropped old android versions that triggered this crash.

This commit remove the ugly jniloader hack.

Here are the list of .so that are now shipped within the APK:
 - c++_shared.so (from the NDK)
 - libvlc.so (all VLC modules, VLC/LibVLC symbols) => link with c++_shared.so
 - libvlcjni.so (all JNI symbols) => link with libvlc.so
 - libmla.so (medialibrary, optional)  => link with libvlc.so and c++_shared.so

The APK size won't change with this commit (the difference of .so sizes is
negligible).