Benchmark: Move screenshots to VLCBenchmark app
As of Android 10, the MediaProjection API previously used to take screenshots
now requires user input to accept permission to take a screenshot at every
restart of the app, which happens for every test on vlc-android.
By using this API in a service in VLCBenchmark, the user is only asked
permission once, when the user starts the benchmark.