Benchmark: Add speed decoding test
This adds a new type of test to the benchmark, like playback, or quality
previously. This test will play a sample and depending on the devices'
ability to play the sample, will slow down, or speed up playback, and
after a few iterations, will converge on a maximum playback speed.