libvlc: VideoHelper: configure VideoTexture correctly
The view to setup can be either mSurfaceVideo or mSurfaceTexture but
mSurfaceVideo was unconditionnaly setup and mSurfaceTexture was never
setup. In addition, the SurfaceTexture was extracted from the
TextureView before it's creation so it was null.

Instead, only setup the SurfaceTexture from the listener call and use
an intermediate variable to configure the correct surface.

fix #1359