Benchmark: fix speed tests on older devices
On most devices, when playback speed goes above 1.0, the metric used to
determine that playback is having trouble will be late frames, with
frame drop when speed is below 1. On some older devices, this behaviour changes,
which leads to an infinite loop because of the wrong metric being used
as a stop condition.
This is a workaround to get speed test to stop on older devices and get
a failure, with a score of zero.

(cherry picked from commit 4faaae49449cfacbab411416250da319bb776f0c)