Benchmark: fix hardware decoding
There was a change with the player restart which is needed when forcing
android-surface vout instead of gles2. This meant that the playlist
manager was restarted instead of the video player. This adds a separate
restart for the playlist manager

(cherry picked from commit b4c108937abc1a1f8335b69222dc0b5e14d00c21)