buildsystem: libvlc: define PKG_CONFIG_LIBDIR
Define PKG_CONFIG_LIBDIR to avoid use pkg-config files from the system.
It didn't happen before because we were using absolute path to the
contribs pkg-config files, so it's not actually fixing a current issue.

We keep PKG_CONFIG_PATH because we must use the same PKG_CONFIG_PATH as
the actual libvlc build, ie. not use anything outside of contribs
currently.