Benchmark: fix playback loop
If repeat type is ONE, playback will stay stuck in a loop.
To fix that, the repeat type is saved, then set to NONE, and at the end
reset to what it was for all benchmark tests that are not speed.

(cherry picked from commit 98316971201e8af87f96993a61d26879f430a9cf)