Bookmarks: fix name update after rename
The adapter update causing the ui update after a rename is triggered by
the observer on the BookmarkModel dataset, by then the dataset has
already changed, and the DiffUtilAdapter will compare the dataset and
the new BookmarkList which are equal. To trigger the name update the
adapter.update must be called before the BookmarkModel.refresh that
actuates the dataset, when dataset and the new bookmark list are still
different.

Closes #2177