Ask for the MANAGE_EXTERNAL_STORAGE permission to existing users
When the old READ_EXTERNAL_STORAGE permission has already been granted
by the user on a previous version of the app, some media may be missing
while they haven't granted the new MANAGE_EXTERNAL_STORAGE permission.
If we detect that the user is in that state, we show a snackbar and let
them grant the new permission.