SubDownload: fix err msg staying at first launch
When starting the dialog for the first time in the app's lifecycle, the
coroutine started search will fail, before succeeding. Because of this,
the errror message is displayed, but never removed once search has been
successful, causing an overlap of the different views. This is fixed by
resetting errorObservable and messageObservable.