MediaParsingService: add delay before service kill
In some discovery cases, the medialibrary will internally switch
isWorking to false, before enqueuing another operation. This means that
the service, that has a listener on isWorking, will stop itself, and wont
get any updates on the operation that the medialibrary enqueued. To
conter this, a delay is added, afterwhich the isWorking value is checked
again before killing the service. If at that point the medialibrary is
working, then the service wont commit suicide, and the ui can get
progress updates.