AspectRatio: fix wrong orientation with locale
When setting a non default locale, the VideoPlayerActivity when in best
fit aspect ratio will not update it's configuration.orientation when
rotating, causing libvlc to assume that the orientation is still the one
used when starting the video. By forcing the orientation to its new
value in onConfigurationChange, this is fixed. Though it doesn't explain
why it has this behaviour in the first place. This is a workaround.

Fixes #2475