AudioProgress: fix audio sometimes being saved
There seems to be a race condition when adding an audio media to history
when it starts, in which it doesnt get the right progress, and hence
getting the progress saved when it shouldn't.
To fix this, a new method was added to media on the medialibrary side,
markAsPlayed to increment playcount and add to history but without
having to call savePosition / Time.