Playlist: fix delete in search
To delete a playlist track requires it's position in the playlist.
When deleting from results of a search it will use the position in the
search results for deletion, and not the position in the playlist,
deleting the wrong track. This just get's the index in the playlist
tracks from media before deleting.

Fixes #2652