VideoPlayer: save media index when leaving app.
On some low performance devices, when leaving the app, the paused
service will be killed, and respawned when the app is restarted.
As the medialist was saved but not the index, resuming would go to
the 1st media in the saved medialist. This saves the media index
also.

Fixes #2705


(cherry picked from commit 72ed253419de5be64ca244e5d932920fa04567a2)