MediaWrapper: add cond to updateMeta for title
If libvlc doesn't have any meta for title it will provide the filename.
This can be problematic with m3u8 #EXTINF that provides a title.
This adds a check when updating the title meta. If the libvlc title
provided is equal to the filename then it will proritize the
medialibrary title. Network share will have null filenames though,
meaning that libvlc will never be chosen, hence the check for directory
type in the condition

Fixes #2722