Revert "Ml: fix discovery-failure after onboarding skip"
This reverts commit 2338bd92a00c8e9ce00109cca5f3bb56f37b4e38.

This commit causes medialibrary in sdcard to be dropped as it permits
the addition of folders as ML devices, which is not what they are.
This causes the medialibrary remove these "devices" when
updatingDevices, hence removing the selected folders.

Fixes #2791


(cherry picked from commit b2e0b54eaf43e3c7e50bfee28514652778ba8bfd)