Ml: fix discovery-failure after onboarding skip
If storage access isn't granted, the medialibrary init will not add the
internal storage as a medialibrary device. After this, it is only
possible to add external devices. So later discovery of internal storage
will fail, not being registered in the medialibrary database.
To fix this, in the late device addition, a check for internal storage
is added.

Fixes #2689


(cherry picked from commit 1e89a2a66cfb91822160644c8e5a46a84cb88acd)