VideoInBackground: add restore tip shown count
Everytime the audio player is created, if a video is playing only audio,
ie a user left the app with video playing in the background, it will
show a snacker tip to restore to video by long pressing the cover.
This should only be shown a few times. Four in this case.