ExternalBrowser: avoid using fd if possible
This is to use the media path directly if possible and avoid using the
fd forwarded to the app from an external browser. Using a fd means that
the app's permission to use the file will disappear which creates issues
when resuming media. Also, opening media from an fd means that the app
won't have access to any meta data.

Fixes #2496, #2861, #2862