Videos: block banning root folder
Since the addition the ban folder feature to the video section through
context actions, it became possible to ban root folder which wasn't
possible from the "Media library folders" view.

This means that once banned, it isn't possible to unban a root folder.
Because of this, it should be blocked at the context action level, by
checking a path being banned with all root paths in the medialibrary.

Fixes #3042