RemoteAcces: fix null media description crash
If for some reason the provider is stuck, the description in the
getProviderContent can be null if the underlaying
MedialibraryItem.description hasn't yet been initialized.

This can happen by triggering the smd auth dialog from the remote
access, leaving it untouched on the phone, and then from the remote
access querying any list of directories, their description is stuck as
long as the smb dialog is still opened.

Fixes #3037