PiP: screen off shouln't set pip to false
In the onStop of the VideoPlayerActivity, isInPiPMode is set to off so
that when reopening the player, it resumes normally.

If the PiP is running, and the screen is turned off, onStop will be
called, and when turning the screen on, the player will resume with
isInPipMode false, but PiP actually not running which triggers the
HUD. Checking in the onStop if the screen is off stops PiP from
being disabled.