CustomPip: fix next loosing video
When running in custom pip, video playback is not in the
videoPlayerActivity. When starting next media, it checks if there is a
video playing, and if (appIsStarted()) which returns false when customPip
is running and the VideoPlayerActivity has been killed.

Because of this, when playing the following media in a playqueue with
custom pip, there no vout. To fix this, this adds a check to
service.isInPiPMode before removing video.