openMediaNoUi: refactor uri to use getMedia(uri)
This is to use getMedia and get a MediaWrapper with meta data from the
medialibrary instead of a bland/empty MediaWrapper.

When opening a media from an other app, if the media wasn't previously
indexed, getMedia will return null, hence the additional use of
getAbstractMediaWrapper(uri). If not the player would just stop without
any warning.