ABRepeat: add seek margin with fast seek
Because fast seek is not precise, it will usually seek before the
requested time when seeking with abrepeat, causing the abrepeat to seek
again, and stucks the player in a loop.
Adding a small margin of time in which the seek is considered acceptable
when fast seek is enabled stops it from looping.