VideoPlayer: fix vob videos restarting at the end
The Medialibrary sometimes doesn't have the media's length,
and won't know that the file has ended.

For some files this causes a finished video to keep it's time and when
played again starts at the end, and leave the player.

This forces the ML to mark the file as finished and reset it's time