APContainer: fix AP not showing after screen off
When playing a video with "Play in background" enabled, if the screen is
switched off, the player will switch to audio, triggering a
ACTION_SHOW_PLAYER scheduled with a 100L delay that has been there since
java.
Because the screen is switched off, onPause is called, triggering a
cancellation of ACTION_SHOW_PLAYER in the LifecyleAwareScheduler,
before it can execute it.
To ensure ACTION_SHOW_PLAYER is executed when switching the screen
back on, a flag is set and checked in onResume to call
showAudioPlayerImpl