VideoPlayer: fix audio playback when leaving pip
When you open pip with play in background, when you quit pip, in the
onStop, it will switch to audio. If you check if system Pip is running
you can stop that, but it will stop playback from running with the screen off,
so check for screen off also.
This means that isInPiPMode will stay true,
even though PiP has been killed a while ago, audio playback in
background is still running.