PlayIndex: fix play-as-audio
By getting a media from the medialibrary from the the currentmedia's uri
to ensure that we have the medialibrary metadata for the file, the new
media from the same uri won't have the flags previously set, including
MEDIA_FORCE_AUDIO which start the media as audio.

To fix this, the flags from the previous mediawrapper are applied to the
new one.

Fixed a regression introduced by ea39df7bb705507c41dd081732588335878142ba