Play as audio: fix next reopening video player
Flags are not applied to the media in mediaList but to the media from
medialibrary, once out of playIndex, it doesn't have any flags anymore.
When changing media, it triggers a doUpdate in the audio player that
triggers a switchToVideo, if the media can switch to video and doesn't
have the MEDIA_FORCE_AUDIO flag, it switches back to video

This applies time from medialibrary to media from mediaList.

This fixes a regression introduced with
ea39df7bb705507c41dd081732588335878142ba
and follow up 50402c904b40b73e2200fc08fe99832f1e7483fa