buildsystem: remove usage of GRADLE_VLC_SRC_DIRS
In libvlcjni VLC_OUT_PATH is set to <vlc>/build-android-<triplet>/ndk and only contains temporary files:
- libvlcjni-modules.c
- libvlcjni-symbols.c
They are not used in the java build.

There is not VLC_OUT_PATH/libs folder after libvlcjni is built.
libvlcjni.so is in libvlcjni/jni/libs which we already handle.

And when building in Android Studio we would add a folder "null" in the srcDir (via libvlcjni/libvlc/build.gradle).