buildsystem: get the NDK path and full version
Otherwise stick to the version from the Dockers and let gradle handle the path.

If we want gradle to use the NDK we used to build libvlc/medialibrary we need
to set the full ndkVersion and the path to that NDK.

The default NDK version 28.2.13676358 correspond to the one we have the current
Docker image.